Add dummies
The first intention was like, why the heck should we add dummies.
After thinking about it I realized if we cant test the contracts,
we easily come into the situation when we want to use a contract,
we have to build those ourselves. So why not do one single job,
but this job well.