Generalize paths by just resolving
The idea initially was we get through a path factory a
path instance. This instance is able to resolve a
path we can work with. So when implemented this contract
we were like this function is to generic. So I separated
it into specific smaller chunks. The problem is that its
too specialize. We have a couple of objects like JsonPath,
ViewPath and many more could follow, so instead of changing
this class with every change, we can easily introduce a new
class which may resolve its path with its own functionality.